HTML5学习项目/yolo/requirements.txt
ultralytics==8.0.196
opencv-python==4.8.1.78
torch>=1.8.0
torchvision>=0.9.0
Pillow>=8.2.0
numpy>=1.21.0
matplotlib>=3.3.0
